src/requirements.txt
Django==1.8.1
MySQL-python==1.2.5
wsgiref==0.1.2
ipython==3.1.0
yolk==0.4.3
pytz==2015.4
python-dateutil==2.4.2
xlwt==1.0.0
djangorestframework==3.6.2
djangorestframework-recursive==0.1.1
django-filter==1.0.2
drf-compound-fields==0.2.2
coreapi==2.3.0
django-rest-swagger==2.1.2
factory_boy==2.8.1
django-rest-assured==0.2.1
Faker==0.7.10